CSCTTV-4140 TTV model update. dim_name.id changed from int to long. Catch up other changes made to SQL DB structure.

+++ aspnetcore/src/DatabaseContext/Entities/BrParticipatesInFundingGroup.cs
@@ -7,7 +7,7 @@ public partial class BrParticipatesInFundingGroup
 {
     public int DimFundingDecisionid { get; set; }
 
-    public int DimNameId { get; set; }
+    public long DimNameId { get; set; }
 
     public int DimOrganizationId { get; set; }
